requirement.txt
tensorflow-gpu==2.4.2
Pillow==8.3.1
qrcode==7.2
